# docs/swagger/swagger.yaml
basePath: /core/v1
definitions:
...
  lakukan-be_internal_modules_crm_auto_reply_rules_dto.CreateAutoReplyRuleRequest:
    properties:
      conditions:
        description: Can be map or struct, will be marshaled to JSONB
      delay_seconds:
        minimum: 0
        type: integer
      description:
        type: string
      is_enabled:
        type: boolean
      message_template:
        minLength: 1
        type: string
      name:
        maxLength: 255
        minLength: 2
        type: string
      priority:
        minimum: 1
        type: integer
      trigger_type:
        enum:
        - first_message
        - outside_business_hours
        - keyword_match
        - no_sales_available
        - high_priority_lead
        - ai_confidence_low
        - conversation_flow
        type: string
    required:
    - message_template
    - name
    - trigger_type
    type: object
  lakukan-be_internal_modules_crm_auto_reply_rules_dto.UpdateAutoReplyRuleRequest:
    properties:
      conditions: {}
      delay_seconds:
        minimum: 0
        type: integer
      description:
        type: string
      is_enabled:
        type: boolean
      message_template:
        minLength: 1
        type: string
      name:
        maxLength: 255
        minLength: 2
        type: string
      priority:
        minimum: 1
        type: integer
      trigger_type:
        enum:
        - first_message
        - outside_business_hours
        - keyword_match
        - no_sales_available
        - high_priority_lead
        - ai_confidence_low
        - conversation_flow
        type: string
    type: object
  lakukan-be_internal_modules_crm_chats_dto.CreateChatMessageRequest:
    properties:
      content:
        type: string
      delivery_status:
        enum:
        - sent
        - delivered
        - read
        - failed
        type: string
      media_mime_type:
        maxLength: 100
        type: string
      media_size_bytes:
        minimum: 0
        type: integer
      media_url:
        maxLength: 500
        type: string
      message_type:
        enum:
        - text
        - image
        - video
        - file
        - voice
        - location
        type: string
      metadata:
        type: string
      sender_id:
        type: string
      sender_name:
        maxLength: 255
        type: string
      sender_type:
        enum:
        - customer
        - sales
        - system
        - auto_reply
        type: string
    required:
    - content
    - message_type
    - sender_type
    type: object
  lakukan-be_internal_modules_crm_chats_dto.CreateChatRequest:
    properties:
      assigned_to_company_user_id:
        type: string
      category:
        enum:
        - hot
        - warm
        - cold
        type: string
      customer_name:
        maxLength: 255
        minLength: 2
        type: string
      email:
        maxLength: 255
        type: string
      first_message:
        allOf:
        - $ref: '#/definitions/lakukan-be_internal_modules_crm_chats_dto.CreateChatMessageRequest'
        description: First message info (optional, for creating chat with initial
          message)
      lead_id:
        type: string
      phone:
        maxLength: 20
        type: string
      platform:
        enum:
        - whatsapp
        - telegram
        - email
        - phone
        type: string
    required:
    - customer_name
    - phone
    - platform
    type: object
...
  lakukan-be_internal_shared_response.Meta:
    properties:
      pagination:
        $ref: '#/definitions/lakukan-be_internal_shared_response.Pagination'
    type: object
  lakukan-be_internal_shared_response.Pagination:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  lakukan-be_internal_shared_response.Response:
    properties:
      data: {}
      errors:
        additionalProperties:
          items:
            type: string
          type: array
        type: object
      message:
        type: string
      meta:
        $ref: '#/definitions/lakukan-be_internal_shared_response.Meta'
    type: object
host: localhost:8080
info:
  contact:
    email: support@lakukan.com
    name: API Support
    url: http://www.lakukan.com/support
  description: Backend API untuk Lakukan - Multi-tenant Business Management Platform
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Lakukan API
  version: "1.0"
paths:
...
  /crm/v1/auto-reply-rules:
    get:
      consumes:
      - application/json
      description: Get list of auto-reply rules with pagination and filters
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      - description: Search by name, description, or message template
        in: query
        name: search
        type: string
      - description: Filter by trigger type
        in: query
        name: trigger_type
        type: string
      - description: Filter by enabled status
        in: query
        name: is_enabled
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
      security:
      - BearerAuth: []
      summary: Get all auto-reply rules
      tags:
      - CRM - Auto Reply Rules
    post:
      consumes:
      - application/json
      description: Create a new auto-reply rule record
      parameters:
      - description: Auto-reply rule data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/lakukan-be_internal_modules_crm_auto_reply_rules_dto.CreateAutoReplyRuleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
      security:
      - BearerAuth: []
      summary: Create new auto-reply rule
      tags:
      - CRM - Auto Reply Rules
  /crm/v1/auto-reply-rules/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete an auto-reply rule
      parameters:
      - description: Auto-reply rule ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
      security:
      - BearerAuth: []
      summary: Delete auto-reply rule
      tags:
      - CRM - Auto Reply Rules
    get:
      consumes:
      - application/json
      description: Get a single auto-reply rule by ID
      parameters:
      - description: Auto-reply rule ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
      security:
      - BearerAuth: []
      summary: Get auto-reply rule by ID
      tags:
      - CRM - Auto Reply Rules
    put:
      consumes:
      - application/json
      description: Update an existing auto-reply rule
      parameters:
      - description: Auto-reply rule ID
        in: path
        name: id
        required: true
        type: string
      - description: Auto-reply rule data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/lakukan-be_internal_modules_crm_auto_reply_rules_dto.UpdateAutoReplyRuleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
      security:
      - BearerAuth: []
      summary: Update auto-reply rule
      tags:
      - CRM - Auto Reply Rules
  /crm/v1/auto-reply-rules/{id}/restore:
    post:
      consumes:
      - application/json
      description: Restore a soft deleted auto-reply rule
      parameters:
      - description: Auto-reply rule ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
      security:
      - BearerAuth: []
      summary: Restore auto-reply rule
      tags:
      - CRM - Auto Reply Rules
  /crm/v1/chats:
    get:
      consumes:
      - application/json
      description: Get list of all chat sessions with pagination and filters
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      - description: Search by customer name, phone, or email
        in: query
        name: search
        type: string
      - description: Filter by platform (whatsapp, telegram, email, phone)
        in: query
        name: platform
        type: string
      - description: Filter by status (active, archived, closed)
        in: query
        name: status
        type: string
      - description: Filter by category (hot, warm, cold)
        in: query
        name: category
        type: string
      - description: Filter by assigned user
        in: query
        name: assigned_to_company_user_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
      security:
      - BearerAuth: []
      summary: Get all chats
      tags:
      - CRM - Chats
    post:
      consumes:
      - application/json
      description: Create a new chat session (for WhatsApp webhook integration)
      parameters:
      - description: Chat data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/lakukan-be_internal_modules_crm_chats_dto.CreateChatRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
      security:
      - BearerAuth: []
      summary: Create new chat
      tags:
      - CRM - Chats
  /crm/v1/chats/{id}:
    get:
      consumes:
      - application/json
      description: Get a single chat with all its messages
      parameters:
      - description: Chat ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lakukan-be_internal_shared_response.Response'
      security:
      - BearerAuth: []
      summary: Get chat detail with messages
      tags:
      - CRM - Chats
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"